✨ (#55) feat: auth store 구현
