+++ src/features/auth/store/authStore.ts
@@ -1,0 +1,30 @@
+import { create } from 'zustand';
+import { persist } from 'zustand/middleware';
+
+// TODO: 백엔드와 협의 후 httpOnly 쿠키 방식으로 전환 권장
+// - 현재: localStorage에 저장 → XSS 취약점 존재 (BARO 구조상 현실적 위험은 낮음)
+// - 개선: 백엔드가 Set-Cookie(HttpOnly, Secure, SameSite=Strict)로 토큰 관리
+//         프론트 authStore 불필요해지고 axios credentials: 'include' 방식으로 전환
+
+interface AuthState {
+  accessToken: string | null;
+  refreshToken: string | null;
+  setAccessToken: (token: string) => void;
+  setTokens: (accessToken: string, refreshToken: string) => void;
+  clearAuth: () => void;
+}
+
+const useAuthStore = create<AuthState>()(
+  persist(
+    (set) => ({
+      accessToken: null,
+      refreshToken: null,
+      setAccessToken: (token) => set({ accessToken: token }),
+      setTokens: (accessToken, refreshToken) => set({ accessToken, refreshToken }),
+      clearAuth: () => set({ accessToken: null, refreshToken: null }),
+    }),
+    { name: 'baro-auth' },
+  ),
+);
+
+export default useAuthStore;
